[modbus_controller] Bound the covered-range join and shrink the offset bias field
The "already covered" branch checked only the upper bound of the range span.
force_new_range sensors sort before plain ones, so iteration is not
address-monotonic: a forced sensor at a high address opened the range and a
following plain sensor at a lower address was absorbed into it with a wrapped
byte offset - its registers were never polled and register-type entities read
out of bounds. The branch now requires the sensor to sit inside the range
(lower and upper bound), never folds into a force-opened range (preserving its
isolation), and only applies to standard-width ranges, since a response_size
register makes interior byte positions unknowable - those fall back to their
own range as before.

shared_offset_bias becomes uint8_t: the old code accumulated into the uint8_t
offset field itself, so this matches its arithmetic exactly and fits the
struct's padding hole instead of growing every SensorItem.

The integration test adds a forced high-address sensor plus a plain low-address
sensor and pins that both poll as separate ranges; the component test config
adds the same shape.

+++ esphome/components/modbus_controller/modbus_controller.h
@@ -152,7 +152,8 @@ class SensorItem {
   uint16_t range_data_offset{0};
   /// Sum of the configured offsets of earlier sensors sharing this register (re-use chains): each
   /// sensor's data historically starts where the previous sensor's offset pointed. 0 normally.
-  uint16_t shared_offset_bias{0};
+  /// uint8_t like offset itself (the old code accumulated into offset directly), and free in padding.
+  uint8_t shared_offset_bias{0};
   uint16_t skip_updates{0};
   std::vector<uint8_t> custom_data{};
   bool force_new_range{false};
